+++ test/test.ml
@@ -11,6 +11,7 @@ let () =
         RFC7515.suite;
         RFC7520.suite;
         RFC7638.suite;
+        RFC8037.suite;
       ]
   in
   match path with Some path -> Junit.to_file report path | None -> ()